# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger ManhDanBlogs",
        "description": "This is a sample Userstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Development Environment"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logs user into the system",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "register a new user - Đang lỗi swagger",
                "operationId": "signup",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "The first name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "The last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "The password_confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "blogs": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "image": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "string"
                    },
                    "description": {
                        "type": "text"
                    },
                    "short_description": {
                        "type": "text"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "categories": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "orders": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "shipping_country": {
                        "type": "string"
                    },
                    "shipping_first_name": {
                        "type": "string"
                    },
                    "shipping_last_name": {
                        "type": "string"
                    },
                    "shipping_address": {
                        "type": "string"
                    },
                    "shipping_optional": {
                        "type": "string"
                    },
                    "shipping_city": {
                        "type": "string"
                    },
                    "shipping_postal_code": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "double"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "orders_detail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "order_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "double"
                    },
                    "amount": {
                        "type": "double"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "products": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "origin_price": {
                        "type": "double"
                    },
                    "sell_price": {
                        "type": "double"
                    },
                    "discount_percent": {
                        "type": "integer"
                    },
                    "discount_price": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "text"
                    },
                    "short_description": {
                        "type": "text"
                    },
                    "size": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "quantity_sell": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "product_comments": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "customer_name": {
                        "type": "string"
                    },
                    "customer_email": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "review_title": {
                        "type": "string"
                    },
                    "body_title": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "user": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "user_wishlist": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "delete_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth API"
        },
        {
            "name": "Admin",
            "description": "Admin api"
        },
        {
            "name": "User",
            "description": "User Api"
        }
    ]
}